Test classes added #20

+++ src/ZigBeeNet/ExtendedPanId.cs
@@ -10,7 +10,8 @@ namespace ZigBeeNet
     {
         public byte[] PanId { get; private set; }
 
-        public ulong Value {
+        public ulong Value
+        {
             get
             {
                 return BitConverter.ToUInt64(PanId, 0);
@@ -18,23 +19,23 @@ namespace ZigBeeNet
         }
 
         /// <summary>
-         /// Default constructor. Creates a PAN Id of 0
-         /// </summary>
+        /// Default constructor. Creates a PAN Id of 0
+        /// </summary>
         public ExtendedPanId()
         {
             PanId = new byte[] { 0, 0, 0, 0, 0, 0, 0, 0 };
         }
 
         /// <summary>
-         /// Create an <see cref="ExtendedPanId"> from a <see cref="BigInteger">
-         ///
-         /// <param name="panId">the panId as a <see cref="BigInteger"></param>
-         /// </summary>
-        public ExtendedPanId(BigInteger panId)
+        /// Create an <see cref="ExtendedPanId"> from a <see cref="BigInteger">
+        ///
+        /// <param name="panId">the panId as a <see cref="long"></param>
+        /// </summary>
+        public ExtendedPanId(long panId)
         {
             PanId = new byte[8];
 
-            long longVal = (long)panId;
+            long longVal = panId;
 
             PanId[0] = (byte)(longVal & 0xff);
             PanId[1] = (byte)((longVal >> 8) & 0xff);
@@ -47,22 +48,22 @@ namespace ZigBeeNet
         }
 
         /// <summary>
-         /// Create an <see cref="ExtendedPanId"> from a <see cref="String"> defined in hexadecimal notation.
-         ///
-         /// <param name="panId">the panId as a <see cref="String"></param>
-         /// </summary>
+        /// Create an <see cref="ExtendedPanId"> from a <see cref="String"> defined in hexadecimal notation.
+        ///
+        /// <param name="panId">the panId as a <see cref="String"></param>
+        /// </summary>
         public ExtendedPanId(string panId)
-            : this(BigInteger.Parse(panId, System.Globalization.NumberStyles.HexNumber))
+            : this(long.Parse(panId, System.Globalization.NumberStyles.HexNumber))
         {
 
         }
 
         /// <summary>
-         /// Create an <see cref="ExtendedPanId"> from an int array
-         ///
-         /// <param name="panId">the panId as an int array. Array length must be 8.</param>
-         /// @throws InvalidParameterException
-         /// </summary>
+        /// Create an <see cref="ExtendedPanId"> from an int array
+        ///
+        /// <param name="panId">the panId as an int array. Array length must be 8.</param>
+        /// @throws InvalidParameterException
+        /// </summary>
         public ExtendedPanId(byte[] panId)
         {
             if (panId == null)
@@ -82,11 +83,11 @@ namespace ZigBeeNet
 
 
         /// <summary>
-         /// Check if the ExtendedPanId is valid. This checks the length of the ID, and checks
-         /// it is not 0000000000000000 or FFFFFFFFFFFFFFFF.
-         ///
-         /// <returns>true if the extended PAN ID is valid</returns>
-         /// </summary>
+        /// Check if the ExtendedPanId is valid. This checks the length of the ID, and checks
+        /// it is not 0000000000000000 or FFFFFFFFFFFFFFFF.
+        ///
+        /// <returns>true if the extended PAN ID is valid</returns>
+        /// </summary>
         public bool IsValid()
         {
             if (PanId == null || PanId.Length != 8)
